CXO-2951 - Universal protocols part 1
 Another batch of changes to make protocols independent of MDW.
 Moved/copied some classes from mdw to protocols.
 Replaced some mdw usages by UPL stuff.
 Added some extra services to find/extract certificates in protocol code
 Removed unused stuff.

+++ upl/com.energyict.mdc.upl.impl/protocolimpl-v2/src/main/java/com/energyict/protocolimplv2/elster/ctr/MTU155/profile/TraceCProfileParser.java
@@ -286,7 +286,7 @@ public class TraceCProfileParser {
      * @return
      */
     private int getIntervalInSeconds() {
-        return getPeriod().getIntervalInSeconds();
+        return (int) getPeriod().getIntervalInSeconds();
     }
 
     /**
